+++ react-client/src/hooks/useFindTrack.tsx
@@ -1,11 +1,13 @@
-import { useState } from 'react';
 import useDebounce from './useDebounce';
 import useSWRInfinite from 'swr/infinite';
-import { findTracks } from '@utils/\bswrFetcher';
+import { findArtists, findTracks } from '@utils/\bswrFetcher';
 
-export function useFindTrack() {
-  const [query, setQuery] = useState('');
-  const debouncedQuery = useDebounce(query, 200);
+interface Prob{
+query:string
+}
+
+export function useFindTrack({ query }:Prob) {
+  const debouncedQuery = useDebounce(query, 100);
   // 트랙 데이터 관리
   const {
     data: trackData,
@@ -16,11 +18,12 @@ export function useFindTrack() {
   } = useSWRInfinite(
     (pageIndex, previousPageData) => {
       if (previousPageData && previousPageData.length === 0) return null; // 마지막 페이지에 도달 시 요청 중지
+      if (query.length === 0) return null;
       const offset = pageIndex * 5; // offset 계산
       return debouncedQuery ? ['tracks', debouncedQuery, offset] : null;
     },
     ([, query, offset]) => findTracks(query, offset),
-    { revalidateOnFocus: false },
+    { revalidateOnFocus: false, dedupingInterval: 300_000 },
   );
 
   async function loadMoreTracks() {
@@ -29,6 +32,38 @@ export function useFindTrack() {
   }
 
   return {
-    setQuery, loadMoreTracks, trackData, trackError, trackIsLoading, setTrackSize, query,
+    loadMoreTracks, trackData, trackError, trackIsLoading, setTrackSize,
+  };
+}
+
+export function useFindArtists({ query }:Prob) {
+  const debouncedQuery = useDebounce(query, 100);
+  const {
+    data: artistData,
+    error: artistError,
+    isLoading: artistIsLoading,
+    size: artistSize,
+    setSize: setArtistSize,
+  } = useSWRInfinite(
+    (pageIndex, previousPageData) => {
+      if (previousPageData && previousPageData.length === 0) return null; // 마지막 페이지에 도달 시 요청 중지
+      if (query.length === 0) return null;
+      const offset = pageIndex * 5; // offset 계산
+      return debouncedQuery ? ['artists', debouncedQuery, offset] : null;
+    },
+    ([, query, offset]) => findArtists(query, offset),
+    {
+      revalidateOnFocus: false, // 페이지로 돌아오거나 브라우저 탭을 다시 포커스할 때 현재 캐시된 데이터를 그대로 사용
+      dedupingInterval: 300_000, // 5분간 동일한 key에대해 재요청 x
+    },
+  );
+
+  async function loadMoreArtists() {
+    const newSize = artistSize + 1;
+    setArtistSize(newSize);
+  }
+
+  return {
+    artistData, artistError, artistIsLoading, setArtistSize, loadMoreArtists,
   };
 }
